requirements.txt
torch==2.9.1
torchvision==0.24.1
Pillow==12.0.0
streamlit==1.51.0
pandas==2.3.3
python-dotenv==1.2.1